refactor(frontend): extract quote bucket section
